Iterated Dynamics : Patched : Apply patch 10 from fractint sources

+++ fractint/dos_help/hc.c
@@ -270,7 +270,7 @@ struct
 int include_stack_top = -1;
 
 
-#define CHK_BUFFER(off) { if ((unsigned)(curr+(off)) - (unsigned)buffer >= (BUFFER_SIZE-1024)) fatal(0,"Buffer overflowed -- Help topic too large."); }
+#define CHK_BUFFER(off) { if ((long)(curr+(off)) - (long)buffer >= (BUFFER_SIZE-1024)) fatal(0,"Buffer overflowed -- Help topic too large."); }
 
 #ifdef __WATCOMC__
 #define putw( x1, x2 )  fprintf( x2, "%c%c", x1&0xFF, x1>>8 );
